rplugin/python3/py_nvim_gpt/requirements.txt
openai
EdgeGPT
googlesearch-python
